Telegram-Bot/requirements_tg.txt
pandas
numpy
matplotlib
seaborn
python-telegram-bot==13.7
pyarrow==14.0.1